+++ packages/app/docs/stories/appWithReduxForm.story.js
@@ -91,7 +91,3 @@ export const simpleFormValidation = () => {
         </RainbowFirebaseApp>
     );
 };
-
-export default {
-    title: 'Modules|App/Stories',
-};