+++ packages/record/docs/stories/RecordDetails.story.js
@@ -67,5 +67,5 @@ export const BasicRecordDetails = () => {
 };
 
 export default {
-    title: 'Modules|Record/Stories/',
+    title: 'Modules/Record/Stories/RecordDetails',
 };