+++ packages/listview/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.23.0](https://github.com/nexxtway/rainbow-modules/compare/v0.22.0...v0.23.0) (2021-01-19)
+
+**Note:** Version bump only for package @rainbow-modules/listview
+
+# [0.22.0](https://github.com/nexxtway/rainbow-modules/compare/v0.21.0...v0.22.0) (2021-01-08)
+
+### Features
+
+-   add DraggableList component to listview module ([#289](https://github.com/nexxtway/rainbow-modules/issues/289)) ([f6a4ee3](https://github.com/nexxtway/rainbow-modules/commit/f6a4ee395c11ccdeb5eed24be01d3912e82a1933))
+
 # [0.21.0](https://github.com/nexxtway/rainbow-modules/compare/v0.20.2...v0.21.0) (2021-01-03)
 
 **Note:** Version bump only for package @rainbow-modules/listview